Add new class SourceException that is the parent exception class for all custom exceptions
