+++ lib/Sabberworm/CSS/Parsing/OutputException.php
@@ -5,17 +5,8 @@ namespace Sabberworm\CSS\Parsing;
 /**
 * Thrown if the CSS parsers attempts to print something invalid
 */
-class OutputException extends \Exception {
-    private $iLineNo;
+class OutputException extends SourceException {
     public function __construct($sMessage, $iLineNo = 0) {
-        $this->$iLineNo = $iLineNo;
-        if (!empty($iLineNo)) {
-            $sMessage .= " [line no: $iLineNo]";
-        }
-        parent::__construct($sMessage);
-    }
-
-    public function getLineNo() {
-        return $this->iLineNo;
+        parent::__construct($sMessage, $iLineNo);
     }
 }
